% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSNPsFromVCF.R
\name{auxGetVcfSource}
\alias{auxGetVcfSource}
\title{auxGetVcfSource}
\usage{
auxGetVcfSource(vcf.source = NULL, vcf.file)
}
\arguments{
\item{vcf.source}{VCF source. Leave NULL to allow the function to recognize it. Otherwise, the function will not try to recognize the source. (Defaults to NULL)}

\item{vcf.file}{VCF file path}
}
\value{
VCF source
}
\description{
Obtains VCF source from a given VCF file path and checks if the source is supported. Auxiliar function used by \code{loadSNPsFromVCF}.
}
