% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVariantsForCNV.R
\name{plotVariantsForCNV}
\alias{plotVariantsForCNV}
\title{plotVariantsForCNV}
\usage{
plotVariantsForCNV(cnvfilter.results, cnv.id, points.cex = 1,
  points.pch = 19, legend.x.pos = 0.08, legend.y.pos = 0.25,
  cnv.zoom.margin = TRUE)
}
\arguments{
\item{cnvfilter.results}{S3 object returned by \code{filterCNVs} function}

\item{cnv.id}{CNV id for which to plot variants}

\item{points.cex}{Points cex (size). (Defaults to 1)}

\item{points.pch}{Points pch (symbol). (Defaults to 19)}

\item{legend.x.pos}{Legend x position. (Defaults to 0.08)}

\item{legend.y.pos}{Legend y position. (Defaults to 0.25)}

\item{cnv.zoom.margin}{If TRUE, the zoom leaves an small margin at both sides of the CNV. False otherwise. (Defaults to TRUE)}
}
\value{
invisibly returns a \code{karyoplot} object
}
\description{
Plots a CNV with all the variants matching it
}
\examples{
# Load CNVs data
cnvs.file <- system.file("extdata", "DECoN.CNVcalls.csv", package = "CNVfilteR", mustWork = TRUE)
cnvs.gr <- loadCNVcalls(cnvs.file = cnvs.file, chr.column = "Chromosome", start.column = "Start", end.column = "End", cnv.column = "CNV.type", sample.column = "Sample")

# Load VCFs data
vcf.files <- c(system.file("extdata", "variants.sample1.vcf.gz", package = "CNVfilteR", mustWork = TRUE),
               system.file("extdata", "variants.sample2.vcf.gz", package = "CNVfilteR", mustWork = TRUE))
vcfs <- loadVCFs(vcf.files, cnvs.gr = cnvs.gr)

# Filter CNVs
results <- filterCNVs(cnvs.gr, vcfs)

# Check CNVs that can be filtered
as.data.frame(results$cnvs[results$cnvs$filter == TRUE])

# Plot one of them
plotVariantsForCNV(results, "3")


}
